requirements.txt
aniso8601==9.0.1
astroid==2.6.6
attrs==20.3.0
bcrypt==3.2.0
black==20.8b1
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
click==8.0.1
codecov==2.1.11
coverage==5.5
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-gzip==0.2
Flask-JWT-Extended==4.1.0
Flask-RESTful==0.3.8
flask-restx==0.5.1
Flask-SQLAlchemy==2.4.1
gevent==21.8.0
greenlet==1.1.2
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.0.0
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==2.11.3
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.11.1
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.10.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.8.2
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.3
pytest-cov==2.12.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
regex==2021.4.4
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.36.7
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
